Moved routes/docs to events.py
